Atualiza projetos Flask, autenticação e documentação

+++ 2-Desenvolvimento com Flask/2 Banco de dados e autenticação/app.py
@@ -1,11 +1,12 @@
 from flask import Flask, request, jsonify
 from database import db
 from models.user import User
-from flask_login import LoginManager, login_user, current_user,logout_user, login_required
+from flask_login import LoginManager, login_user, current_user, logout_user, login_required
+import bcrypt
 
 app = Flask(__name__)
 app.config["SECRET_KEY"] = 'your_secret_key'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
+app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://root:admin123@127.0.0.1:3306/flask-crud'
 
 login_manager = LoginManager()
 db.init_app(app)
@@ -15,7 +16,7 @@ login_manager.login_view = 'login'
 
 @login_manager.user_loader
 def load_user(user_id):
-    return User.query.get(user_id)
+    return db.session.get(User, user_id)
 
 @app.route("/login", methods=["POST"])
 def login():
@@ -26,7 +27,7 @@ def login():
     if username and password:
 
         user = User.query.filter_by(username=username).first()
-        if user and user.password == password:
+        if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
             login_user(user)
             print(current_user.is_authenticated)
             return jsonify({'message':'Usuário Logado!'})
@@ -44,9 +45,11 @@ def create_user():
     data = request.get_json()
     username = data.get('username')
     password = data.get('password')
+    role = data.get('role')
 
     if username and password:
-        user = User(username=username, password=password)
+        hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+        user = User(username=username, password=hashed_password, role=role)
         db.session.add(user)
         db.session.commit()
         return jsonify({'message':'Usuário Cadastrado!'})
@@ -56,7 +59,7 @@ def create_user():
 @app.route("/user/<int:user_id>", methods=['GET'])
 @login_required
 def read_user(user_id):
-    user = User.query.get(user_id)
+    user = db.session.get(User, user_id)
 
     if user:
         return {"username": user.username}
@@ -67,11 +70,14 @@ def read_user(user_id):
 @login_required
 def update_user(user_id):
     data = request.get_json()
-    user =  User.query.get(user_id)
+    user = db.session.get(User, user_id)
     password = data.get("password")
 
+    if user_id != current_user.id and current_user.role == 'user':
+        return jsonify({"message":"Operação não permitida!"}), 403
+
     if user and password:
-        user.password = password
+        user.password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
         db.session.commit()
         return jsonify({"message":"senha atualizada com sucesso"})
     
@@ -80,15 +86,18 @@ def update_user(user_id):
 @app.route("/user/<int:user_id>", methods=['DELETE'])
 @login_required
 def delete_user(user_id):
-    user =  User.query.get(user_id)
+    user = db.session.get(User, user_id)
 
+    if current_user.role != 'admin':
+        return jsonify({"message":"Operação não permitida!"}), 403
+    
     if user_id == current_user.id:
         return jsonify({"message":"Deleção não permitida!"}), 403
 
     if user:
         db.session.delete(user)
         db.session.commit()
-        return jsonify({"message": f"Usuário {user_id} atualizada com sucesso"})
+        return jsonify({"message": f"Usuário {user_id} deletado com sucesso"})
     
     return jsonify({'message':'Usuário não encontrado'}), 404
